[fork-join] fix merge queue and nop instructions
The case when a queue instruction is merged with a NOP instruction
was not handled properly. This case can be hit when GS cuts the graph
somewhere on a fork-join path, so we need to buffer the fork-join
with queues.

Let's say that the previous instruction added NOP buffer_0_1_2
between nodes A and B. In the next iteration of epoch balancing,
GS cuts the graph somewhere on that fork-join and we need to
add a buffering queue on the path between A and B.

The new instruction will have nodes A and buffer_0_1_2 as src/dest pair.
So, to properly merge these instructions, we need to create a new
queue instruction with nodes A and B as src/dest pair, and use it
as a replacement for the old (NOP) instruction.

(cherry picked from commit bfd97cdffd9c5d10f2327e230baeaaaea577f0be)